refactor init functions
